test: Customize test timeouts
