+++ test/support/support.go
@@ -17,17 +17,52 @@ limitations under the License.
 package support
 
 import (
+	"fmt"
+	"os"
 	"time"
+
+	"github.com/onsi/gomega"
+	"github.com/onsi/gomega/format"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-const (
+var (
 	TestTimeoutShort  = 1 * time.Minute
 	TestTimeoutMedium = 2 * time.Minute
 	TestTimeoutLong   = 5 * time.Minute
-)
 
-var (
 	ApplyOptions = metav1.ApplyOptions{FieldManager: "codeflare-test", Force: true}
 )
+
+func init() {
+	if value, ok := os.LookupEnv("CODEFLARE_TEST_TIMEOUT_SHORT"); ok {
+		if duration, err := time.ParseDuration(value); err == nil {
+			TestTimeoutShort = duration
+		} else {
+			fmt.Printf("Error parsing CODEFLARE_TEST_TIMEOUT_SHORT. Using default value: %s", TestTimeoutShort)
+		}
+	}
+	if value, ok := os.LookupEnv("CODEFLARE_TEST_TIMEOUT_MEDIUM"); ok {
+		if duration, err := time.ParseDuration(value); err == nil {
+			TestTimeoutMedium = duration
+		} else {
+			fmt.Printf("Error parsing CODEFLARE_TEST_TIMEOUT_MEDIUM. Using default value: %s", TestTimeoutMedium)
+		}
+	}
+	if value, ok := os.LookupEnv("CODEFLARE_TEST_TIMEOUT_LONG"); ok {
+		if duration, err := time.ParseDuration(value); err == nil {
+			TestTimeoutLong = duration
+		} else {
+			fmt.Printf("Error parsing CODEFLARE_TEST_TIMEOUT_LONG. Using default value: %s", TestTimeoutLong)
+		}
+	}
+
+	// Gomega settings
+	gomega.SetDefaultEventuallyTimeout(TestTimeoutShort)
+	gomega.SetDefaultEventuallyPollingInterval(1 * time.Second)
+	gomega.SetDefaultConsistentlyDuration(30 * time.Second)
+	gomega.SetDefaultConsistentlyPollingInterval(1 * time.Second)
+	// Disable object truncation on test results
+	format.MaxLength = 0
+}
